re-organize helpers and util
